implement delete action

+++ trunk/40_implementation/LIMO/LIMO_view/src/main/java/nl/fontys/sofa/limo/view/factory/ProcedureCategoryChildFactory.java
@@ -5,7 +5,6 @@ import java.beans.PropertyChangeEvent;
 import java.util.Collection;
 import java.util.List;
 import nl.fontys.sofa.limo.api.service.provider.ProcedureCategoryService;
-import nl.fontys.sofa.limo.domain.BaseEntity;
 import nl.fontys.sofa.limo.domain.component.procedure.ProcedureCategory;
 import nl.fontys.sofa.limo.view.node.ProcedureCategoryNode;
 import org.openide.nodes.BeanNode;
@@ -23,7 +22,8 @@ import org.openide.util.LookupListener;
 import org.openide.util.Utilities;
 
 /**
- * Factory responsible for creating the CostCategoryNode children.
+ * Factory responsible for creating the ProcedureCategory children. It listens
+ * to changes in the service layer and in the nodes.
  *
  * @author Sebastiaan Heijmann
  */
@@ -60,40 +60,31 @@ public class ProcedureCategoryChildFactory extends ChildFactory<ProcedureCategor
 
 	@Override
 	public void resultChanged(LookupEvent le) {
-		Lookup.Result res = (Lookup.Result) le.getSource();
-        Collection instances = res.allInstances();
-
-        if (!instances.isEmpty()) {
-			refresh(true);
-        }
-	}
-
-	@Override
-	public void childrenAdded(NodeMemberEvent nme) {
-		throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-	}
-
-	@Override
-	public void childrenRemoved(NodeMemberEvent nme) {
-		throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-	}
-
-	@Override
-	public void childrenReordered(NodeReorderEvent nre) {
-		throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+		refresh(true);
 	}
 
 	@Override
 	public void nodeDestroyed(NodeEvent ne) {
 		Node node = ne.getNode();
-		Lookup result = node.getLookup();
-		ProcedureCategory selected = Utilities.actionsGlobalContext().lookup(ProcedureCategory.class);
-		ProcedureCategory pc = (ProcedureCategory) node.getLookup().lookup(ProcedureCategory.class);
-		service.delete(pc);
+		ProcedureCategory pc =
+				(ProcedureCategory) node.getLookup().lookup(ProcedureCategory.class);
+
+		Lookup.Result result = Utilities.actionsGlobalContext().lookupResult (ProcedureCategory.class);
+		Collection<ProcedureCategory> selectedBeans = result.allInstances();
+		for(ProcedureCategory bean : selectedBeans){
+			if(bean == pc){
+				service.delete(pc);
+			}	
+		}
 		refresh(true);
 	}
 
 	@Override
-	public void propertyChange(PropertyChangeEvent pce) {
-	}
+	public void childrenAdded(NodeMemberEvent nme) {}
+	@Override
+	public void childrenRemoved(NodeMemberEvent nme) {}
+	@Override
+	public void childrenReordered(NodeReorderEvent nre) {}
+	@Override
+	public void propertyChange(PropertyChangeEvent pce) {}
 }
